requirements.txt
# Core dependencies
pypdf>=6.1.3
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
statsmodels>=0.14.0

# WebSocket and HTTP
websocket-client>=1.6.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0
plotly>=5.17.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Additional utilities
python-dotenv>=1.0.0
pytz>=2023.3


